+++ src/TitleForm/TitleFormContainer.js
@@ -3,7 +3,10 @@ import PropTypes from 'prop-types';
 import ReactRouterPropTypes from 'react-router-prop-types';
 
 import { stripesConnect } from '@folio/stripes/core';
-import { identifierTypesManifest } from '@folio/stripes-acq-components';
+import {
+  identifierTypesManifest,
+  useShowCallout,
+} from '@folio/stripes-acq-components';
 
 import { titlesResource } from '../common/resources';
 import TitleForm from './TitleForm';
@@ -11,6 +14,7 @@ import TitleForm from './TitleForm';
 function TitleFormContainer({ history, match, mutator }) {
   const titleId = match.params.id;
   const [identifierTypes, setIdentifierTypes] = useState();
+  const showCallout = useShowCallout();
 
   useEffect(() => {
     mutator.identifierTypes.GET()
@@ -24,15 +28,30 @@ function TitleFormContainer({ history, match, mutator }) {
     [history],
   );
   const onSubmit = useCallback(
-    // eslint-disable-next-line no-unused-vars
     ({ poLineNumber, ...newTitle }) => {
       return mutator.titles.POST(newTitle)
         .then(() => {
+          showCallout({
+            messageId: 'ui-receiving.title.actions.save.success',
+            type: 'success',
+            values: {
+              title: newTitle.title,
+              poLineNumber,
+            },
+          });
           setTimeout(() => history.push('/receiving'));
-        });
+        })
+        .catch(() => showCallout({
+          messageId: 'ui-receiving.title.actions.save.error',
+          type: 'error',
+          values: {
+            title: newTitle.title,
+            poLineNumber,
+          },
+        }));
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [history, titleId],
+    [history, showCallout, titleId],
   );
 
   if (!identifierTypes) {