+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.0.0](https://github.com/folio-org/ui-receiving/tree/v1.0.0) (in progress)
 
 ### Stories
+* [UIREC-34](https://issues.folio.org/browse/UIREC-34) Add success message for create title
 * [UIREC-28](https://issues.folio.org/browse/UIREC-28) Add Title for receiving/check-in
 * [UIREC-27](https://issues.folio.org/browse/UIREC-27) View title in receiving area
 * [UIREC-26](https://issues.folio.org/browse/UIREC-26) Create three pane layout for receiving